Refactor out LogoutButton and match style of other header buttons
